drf-intro/simple_crud/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-debug-toolbar==3.2
djangorestframework==3.12.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
